Исправляет ошибку в тесте FilmsList
